Refactor tests into isolated modules
