+++ make/tests/maths.toml
@@ -6,8 +6,8 @@ dependencies = [
 
 [tasks.test-maths-default]
 command = "cargo"
-args = ["test", "--workspace", "--no-default-features", "--features=maths", "maths", "--", "--skip", "generated"]
+args = ["test", "--workspace", "--no-default-features", "--features=maths", "maths"]
 
 [tasks.test-maths-nopanic]
 command = "cargo"
-args = ["test", "--workspace", "--no-default-features", "--features=maths-nopanic", "maths", "--", "--skip", "generated"]
+args = ["test", "--workspace", "--no-default-features", "--features=maths-nopanic", "maths"]
