+++ make/tests/misc.toml
@@ -19,23 +19,23 @@ args = ["target", "add", "x86_64-unknown-none"]
 
 [tasks.test-proptest]
 command = "cargo"
-args = ["test", "--workspace", "--no-default-features", "--features=proptest", "proptest_tests", "--", "--skip", "generated"]
+args = ["test", "--workspace", "--no-default-features", "--features=proptest", "--test", "proptest_tests"]
 
 [tasks.test-rust-fuzz]
 command = "cargo"
-args = ["test", "--workspace", "--no-default-features", "--features=rust-fuzz", "rust_fuzz_tests", "--", "--skip", "generated"]
+args = ["test", "--workspace", "--no-default-features", "--features=rust-fuzz", "--test", "rust_fuzz_tests"]
 
 [tasks.test-rocket-traits]
 command = "cargo"
-args = ["test", "--workspace", "--features=rocket-0_5-traits", "rocket_tests"]
+args = ["test", "--workspace", "--features=rocket-0_5-traits", "--test", "rocket_tests"]
 
 [tasks.test-borsh]
 command = "cargo"
-args = ["test", "--workspace", "--features=borsh", "borsh_tests", "--", "--skip", "generated"]
+args = ["test", "--workspace", "--features=borsh", "--test", "borsh_tests"]
 
 [tasks.test-ndarray]
 command = "cargo"
-args = ["test", "--workspace", "--features=ndarray-0_16", "nd_array_tests", "--", "--skip", "generated"]
+args = ["test", "--workspace", "--features=ndarray-0_16", "--test", "ndarray_tests"]
 
 [tasks.test-rand]
 dependencies = [
@@ -44,4 +44,4 @@ dependencies = [
 
 [tasks.test-rand-0_9]
 command = "cargo"
-args = ["test", "--workspace", "--features=rand-0_9", "rand_tests", "--", "--skip", "generated"]
+args = ["test", "--workspace", "--features=rand-0_9", "rand_tests"]
