+++ tests/decimal_tests/arithmetic.rs
@@ -1,0 +1,449 @@
+use core::str::FromStr;
+use num_traits::ToPrimitive;
+use rust_decimal::Decimal;
+
+// Negation
+#[test]
+fn it_negates_decimals() {
+    fn neg(a: &str, b: &str) {
+        let a = Decimal::from_str(a).unwrap();
+        let result = -a;
+        assert_eq!(b, result.to_string(), "- {a}");
+    }
+
+    let tests = &[
+        ("1", "-1"),
+        ("2", "-2"),
+        ("2454495034", "-2454495034"),
+        ("0.1", "-0.1"),
+        ("11.815126050420168067226890757", "-11.815126050420168067226890757"),
+    ];
+
+    for &(a, b) in tests {
+        neg(a, b);
+        neg(b, a);
+    }
+}
+
+// Addition
+
+#[test]
+fn it_can_add_simple() {
+    // This is the most basic test for addition, intended largely for micro-optimization.
+    let two = Decimal::ONE + Decimal::ONE;
+    assert_eq!(two.to_u32(), Some(2));
+}
+
+#[test]
+fn it_adds_decimals() {
+    fn add(a: &str, b: &str, c: &str) {
+        let a = Decimal::from_str(a).unwrap();
+        let b = Decimal::from_str(b).unwrap();
+        let result = a + b;
+        assert_eq!(c, result.to_string(), "{a} + {b}");
+        let result = b + a;
+        assert_eq!(c, result.to_string(), "{b} + {a}");
+    }
+
+    let tests = &[
+        ("0", "0", "0"),
+        ("0", "-0", "0"),
+        ("-0", "0", "0"),
+        ("-0", "-0", "0"),
+        ("2", "3", "5"),
+        ("2454495034", "3451204593", "5905699627"),
+        ("24544.95034", ".3451204593", "24545.2954604593"),
+        (".1", ".1", "0.2"),
+        (".10", ".1", "0.20"),
+        (".1", "-.1", "0.0"),
+        ("0", "1.001", "1.001"),
+        ("2", "-3", "-1"),
+        ("-2", "3", "1"),
+        ("-2", "-3", "-5"),
+        ("3", "-2", "1"),
+        ("-3", "2", "-1"),
+        ("1.234", "2.4567", "3.6907"),
+        (
+            "11.815126050420168067226890757",
+            "0.6386554621848739495798319328",
+            "12.453781512605042016806722690",
+        ),
+        (
+            "-11.815126050420168067226890757",
+            "0.6386554621848739495798319328",
+            "-11.176470588235294117647058824",
+        ),
+        (
+            "11.815126050420168067226890757",
+            "-0.6386554621848739495798319328",
+            "11.176470588235294117647058824",
+        ),
+        (
+            "-11.815126050420168067226890757",
+            "-0.6386554621848739495798319328",
+            "-12.453781512605042016806722690",
+        ),
+        (
+            "11815126050420168067226890757",
+            "0.4386554621848739495798319328",
+            "11815126050420168067226890757",
+        ),
+        (
+            "-11815126050420168067226890757",
+            "0.4386554621848739495798319328",
+            "-11815126050420168067226890757",
+        ),
+        (
+            "11815126050420168067226890757",
+            "-0.4386554621848739495798319328",
+            "11815126050420168067226890757",
+        ),
+        (
+            "-11815126050420168067226890757",
+            "-0.4386554621848739495798319328",
+            "-11815126050420168067226890757",
+        ),
+        (
+            "0.0872727272727272727272727272",
+            "843.65000000",
+            "843.7372727272727272727272727",
+        ),
+        (
+            "7314.6229858868828353570724702",
+            "1000",
+            // Overflow causes this to round
+            "8314.622985886882835357072470",
+        ),
+        (
+            "108053.27500000000000000000000",
+            "0.00000000000000000000000",
+            "108053.27500000000000000000000",
+        ),
+        (
+            "108053.27500000000000000000000",
+            // This zero value has too high precision and will be trimmed
+            "0.000000000000000000000000",
+            "108053.27500000000000000000000",
+        ),
+        (
+            "108053.27500000000000000000000",
+            // This value has too high precision and will be rounded
+            "0.000000000000000000000001",
+            "108053.27500000000000000000000",
+        ),
+        (
+            "108053.27500000000000000000000",
+            // This value has too high precision and will be rounded
+            "0.000000000000000000000005",
+            "108053.27500000000000000000000",
+        ),
+        (
+            "8097370036018690744.2590371109596744091",
+            "3807285637671831400.15346897797550749555",
+            "11904655673690522144.412506089",
+        ),
+    ];
+    for &(a, b, c) in tests {
+        add(a, b, c);
+    }
+}
+
+#[test]
+fn it_can_addassign() {
+    let mut a = Decimal::from_str("1.01").unwrap();
+    let b = Decimal::from_str("0.99").unwrap();
+    a += b;
+    assert_eq!("2.00", a.to_string());
+
+    a += &b;
+    assert_eq!("2.99", a.to_string());
+
+    let mut c = &mut a;
+    c += b;
+    assert_eq!("3.98", a.to_string());
+
+    let mut c = &mut a;
+    c += &b;
+    assert_eq!("4.97", a.to_string());
+}
+
+// Subtraction
+
+#[test]
+fn it_subtracts_decimals() {
+    fn sub(a: &str, b: &str, c: &str) {
+        let a = Decimal::from_str(a).unwrap();
+        let b = Decimal::from_str(b).unwrap();
+        let result = a - b;
+        assert_eq!(c, result.to_string(), "{a} - {b}");
+    }
+
+    let tests = &[
+        ("0", "0", "0"),
+        ("0", "-0", "0"),
+        ("-0", "0", "0"),
+        ("-0", "-0", "0"),
+        ("2", "3", "-1"),
+        ("3451204593", "2323322332", "1127882261"),
+        ("24544.95034", ".3451204593", "24544.6052195407"),
+        (".1", ".1", "0.0"),
+        (".1", "-.1", "0.2"),
+        ("1.001", "0", "1.001"),
+        ("2", "-3", "5"),
+        ("-2", "3", "-5"),
+        ("-2", "-3", "1"),
+        ("3", "-2", "5"),
+        ("-3", "2", "-5"),
+        ("1.234", "2.4567", "-1.2227"),
+        ("844.13000000", "843.65000000", "0.48000000"),
+        ("79228162514264337593543950335", "79228162514264337593543950335", "0"), // 0xFFFF_FFFF_FFFF_FFFF_FFF_FFFF - 0xFFFF_FFFF_FFFF_FFFF_FFF_FFFF
+        ("79228162514264337593543950335", "0", "79228162514264337593543950335"),
+        ("79228162514264337593543950335", "79228162514264337593543950333", "2"),
+        ("4951760157141521099596496896", "1", "4951760157141521099596496895"), // 0x1000_0000_0000_0000_0000_0000 - 1 = 0x0FFF_FFFF_FFFF_FFFF_FFF_FFFF
+        ("79228162514264337593543950334", "79228162514264337593543950335", "-1"),
+        ("1", "4951760157141521099596496895", "-4951760157141521099596496894"),
+        ("18446744073709551615", "-18446744073709551615", "36893488147419103230"), // 0xFFFF_FFFF_FFFF_FFFF - -0xFFFF_FFFF_FFFF_FFFF
+    ];
+    for &(a, b, c) in tests {
+        sub(a, b, c);
+    }
+}
+
+#[test]
+fn it_can_subassign() {
+    let mut a = Decimal::from_str("1.01").unwrap();
+    let b = Decimal::from_str("0.51").unwrap();
+    a -= b;
+    assert_eq!("0.50", a.to_string());
+
+    a -= &b;
+    assert_eq!("-0.01", a.to_string());
+
+    let mut c = &mut a;
+    c -= b;
+    assert_eq!("-0.52", a.to_string());
+
+    let mut c = &mut a;
+    c -= &b;
+    assert_eq!("-1.03", a.to_string());
+}
+
+// Multiplication
+
+#[test]
+fn it_multiplies_decimals() {
+    fn mul(a: &str, b: &str, c: &str) {
+        let a = Decimal::from_str(a).unwrap();
+        let b = Decimal::from_str(b).unwrap();
+        let result = a * b;
+        assert_eq!(c, result.to_string(), "{a} * {b}");
+        let result = b * a;
+        assert_eq!(c, result.to_string(), "{b} * {a}");
+    }
+
+    let tests = &[
+        ("2", "3", "6"),
+        ("2454495034", "3451204593", "8470964534836491162"),
+        ("24544.95034", ".3451204593", "8470.964534836491162"),
+        (".1", ".1", "0.01"),
+        ("0", "1.001", "0"),
+        ("2", "-3", "-6"),
+        ("-2", "3", "-6"),
+        ("-2", "-3", "6"),
+        ("1", "2.01", "2.01"),
+        ("1.0", "2.01", "2.010"), // Scale is always additive
+        (
+            "0.00000000000000001",
+            "0.00000000000000001",
+            "0.0000000000000000000000000000",
+        ),
+        ("0.0000000000000000000000000001", "0.0000000000000000000000000001", "0"),
+        (
+            "0.6386554621848739495798319328",
+            "11.815126050420168067226890757",
+            "7.5457947885036367488171739292",
+        ),
+        (
+            "2123456789012345678901234567.8",
+            "11.815126050420168067226890757",
+            "25088909624801327937270048761",
+        ),
+        (
+            "2123456789012345678901234567.8",
+            "-11.815126050420168067226890757",
+            "-25088909624801327937270048761",
+        ),
+        (
+            "2.1234567890123456789012345678",
+            "2.1234567890123456789012345678",
+            "4.5090687348026215523554336227",
+        ),
+        (
+            "0.48000000",
+            "0.1818181818181818181818181818",
+            "0.0872727272727272727272727273",
+        ),
+    ];
+    for &(a, b, c) in tests {
+        mul(a, b, c);
+    }
+}
+
+#[test]
+#[should_panic(expected = "Multiplication overflowed")]
+fn it_panics_when_multiply_with_overflow() {
+    let a = Decimal::from_str("2000000000000000000001").unwrap();
+    let b = Decimal::from_str("3000000000000000000001").unwrap();
+    let _ = a * b;
+}
+
+#[test]
+fn it_can_mulassign() {
+    let mut a = Decimal::from_str("1.25").unwrap();
+    let b = Decimal::from_str("0.01").unwrap();
+
+    a *= b;
+    assert_eq!("0.0125", a.to_string());
+
+    a *= &b;
+    assert_eq!("0.000125", a.to_string());
+
+    let mut c = &mut a;
+    c *= b;
+    assert_eq!("0.00000125", a.to_string());
+
+    let mut c = &mut a;
+    c *= &b;
+    assert_eq!("0.0000000125", a.to_string());
+}
+
+// Division
+
+#[test]
+fn it_divides_decimals() {
+    fn div(a: &str, b: &str, c: &str) {
+        let a = Decimal::from_str(a).unwrap();
+        let b = Decimal::from_str(b).unwrap();
+        let result = a / b;
+        assert_eq!(c, result.to_string(), "{a} / {b}");
+    }
+
+    let tests = &[
+        ("6", "3", "2"),
+        ("10", "2", "5"),
+        ("2.2", "1.1", "2"),
+        ("-2.2", "-1.1", "2"),
+        ("12.88", "5.6", "2.3"),
+        ("1023427554493", "43432632", "23563.562864276795382789603909"),
+        ("10000", "3", "3333.3333333333333333333333333"),
+        ("2", "3", "0.6666666666666666666666666667"),
+        ("1", "3", "0.3333333333333333333333333333"),
+        ("-2", "3", "-0.6666666666666666666666666667"),
+        ("2", "-3", "-0.6666666666666666666666666667"),
+        ("-2", "-3", "0.6666666666666666666666666667"),
+        ("1234.5678", "0.1234567890123456", "9999.99926999999982999953127"),
+        ("1234.567890123456789012345678", "1.234567890123456789012345678", "1000"),
+        (
+            "32.91625929004387114334488",
+            "3.27629537",
+            "10.046792359274942644546996384",
+        ),
+        ("5000", "1000.26957490549", "4.9986524887277738570721416846"),
+        ("6142.6941216127122745222131114", "2", "3071.3470608063561372611065557"),
+        (
+            "3071.3470608063561372611065557",
+            "1228.87000756",
+            "2.4993262443638869285360708423",
+        ),
+        (
+            "590.3274854004009467754255123",
+            "53.68997202826239",
+            "10.995116277759516850521689988",
+        ),
+    ];
+    for &(a, b, c) in tests {
+        div(a, b, c);
+    }
+}
+
+#[test]
+#[should_panic(expected = "Division by zero")]
+fn it_can_divide_by_zero() {
+    let a = Decimal::from_str("2").unwrap();
+    let _ = a / Decimal::ZERO;
+}
+
+#[test]
+fn it_can_divassign() {
+    let mut a = Decimal::from_str("1.25").unwrap();
+    let b = Decimal::from_str("0.01").unwrap();
+
+    a /= b;
+    assert_eq!("125", a.to_string());
+
+    a /= &b;
+    assert_eq!("12500", a.to_string());
+
+    let mut c = &mut a;
+    c /= b;
+    assert_eq!("1250000", a.to_string());
+
+    let mut c = &mut a;
+    c /= &b;
+    assert_eq!("125000000", a.to_string());
+}
+
+// Modulus and Remainder are not the same thing!
+// https://math.stackexchange.com/q/801962/82277
+
+#[test]
+fn it_rems_decimals() {
+    fn rem(a: &str, b: &str, c: &str) {
+        let a = Decimal::from_str(a).unwrap();
+        let b = Decimal::from_str(b).unwrap();
+        // a = qb + r
+        let result = a % b;
+        assert_eq!(c, result.to_string(), "{a} % {b}");
+    }
+
+    let tests = &[
+        ("2", "3", "2"),
+        ("-2", "3", "-2"),
+        ("2", "-3", "2"),
+        ("-2", "-3", "-2"),
+        ("6", "3", "0"),
+        ("42.2", "11.9", "6.5"),
+        ("2.1", "3", "2.1"),
+        ("2", "3.1", "2"),
+        ("2.0", "3.1", "2.0"),
+        ("4", "3.1", "0.9"),
+        ("2", "2", "0"),
+        ("2", "-2", "0"),
+        // Legacy keeps sign from lhs operand
+        ("-2", "2", "0"),
+        ("-2", "-2", "0"),
+    ];
+    for &(a, b, c) in tests {
+        rem(a, b, c);
+    }
+}
+
+#[test]
+fn it_can_remassign() {
+    let mut a = Decimal::from_str("5").unwrap();
+    let b = Decimal::from_str("2").unwrap();
+
+    a %= b;
+    assert_eq!("1", a.to_string());
+
+    a %= &b;
+    assert_eq!("1", a.to_string());
+
+    let mut c = &mut a;
+    c %= b;
+    assert_eq!("1", a.to_string());
+
+    let mut c = &mut a;
+    c %= &b;
+    assert_eq!("1", a.to_string());
+}
