+++ tests/decimal_tests/comparison.rs
@@ -1,0 +1,1331 @@
+use core::cmp::Ordering::*;
+use core::str::FromStr;
+use rust_decimal::Decimal;
+
+#[test]
+fn it_eqs_decimals() {
+    fn eq(a: &str, b: &str, c: bool) {
+        let a = Decimal::from_str(a).unwrap();
+        let b = Decimal::from_str(b).unwrap();
+        assert_eq!(c, a.eq(&b), "{a} == {b}");
+        assert_eq!(c, b.eq(&a), "{b} == {a}");
+    }
+
+    let tests = &[
+        ("1", "1", true),
+        ("1", "-1", false),
+        ("1", "1.00", true),
+        ("1.2345000000000", "1.2345", true),
+        ("1.0000000000000000000000000000", "1.0000000000000000000000000000", true),
+        (
+            "1.0000000000000000000000000001",
+            "1.0000000000000000000000000000",
+            false,
+        ),
+    ];
+    for &(a, b, c) in tests {
+        eq(a, b, c);
+    }
+}
+
+#[test]
+fn it_cmps_decimals() {
+    fn cmp(a: &str, b: &str, c: core::cmp::Ordering) {
+        let a = Decimal::from_str(a).unwrap();
+        let b = Decimal::from_str(b).unwrap();
+        assert_eq!(
+            c,
+            a.cmp(&b),
+            "{} {} {}",
+            a,
+            match c {
+                Less => "<",
+                Equal => "==",
+                Greater => ">",
+            },
+            b
+        );
+    }
+
+    let tests = &[
+        ("1", "1", Equal),
+        ("1", "-1", Greater),
+        ("1", "1.00", Equal),
+        ("1.2345000000000", "1.2345", Equal),
+        (
+            "1.0000000000000000000000000001",
+            "1.0000000000000000000000000000",
+            Greater,
+        ),
+        ("1.0000000000000000000000000000", "1.0000000000000000000000000001", Less),
+        ("-1", "100", Less),
+        ("-100", "1", Less),
+        ("0", "0.5", Less),
+        ("0.5", "0", Greater),
+        ("100", "0.0098", Greater),
+        ("1000000000000000", "999000000000000.0001", Greater),
+        ("2.0001", "2.0001", Equal),
+        (
+            "11.815126050420168067226890757",
+            "0.6386554621848739495798319328",
+            Greater,
+        ),
+        ("0.6386554621848739495798319328", "11.815126050420168067226890757", Less),
+        ("-0.5", "-0.01", Less),
+        ("-0.5", "-0.1", Less),
+        ("-0.01", "-0.5", Greater),
+        ("-0.1", "-0.5", Greater),
+        // 000 equality
+        ("0.00000000", "0.00000000", Equal),
+        // 000 000 same scale
+        ("0.00000000", "0.00000000", Equal),
+        ("-0.00000000", "0.00000000", Equal),
+        ("0.00000000", "-0.00000000", Equal),
+        ("-0.00000000", "-0.00000000", Equal),
+        // 000 000 different scale
+        ("0.000000000", "0.00000000000000000000000", Equal),
+        ("-0.000000000", "0.00000000000000000000000", Equal),
+        ("0.000000000", "-0.00000000000000000000000", Equal),
+        ("-0.000000000", "-0.00000000000000000000000", Equal),
+        // 000 100 same scale
+        ("0.00000000", "6.56792910", Less),
+        ("-0.00000000", "6.56792910", Less),
+        ("0.00000000", "-6.56792910", Greater),
+        ("-0.00000000", "-6.56792910", Greater),
+        // 000 100 different scale
+        ("0.0000000000000000000", "0.00000000000000001916236746", Less),
+        ("-0.0000000000000000000", "0.00000000000000001916236746", Less),
+        ("0.0000000000000000000", "-0.00000000000000001916236746", Greater),
+        ("-0.0000000000000000000", "-0.00000000000000001916236746", Greater),
+        // 000 010 same scale
+        ("0.00000000", "49037796231.72571136", Less),
+        ("-0.00000000", "49037796231.72571136", Less),
+        ("0.00000000", "-49037796231.72571136", Greater),
+        ("-0.00000000", "-49037796231.72571136", Greater),
+        // 000 010 different scale
+        ("0", "14459264155.12895488", Less),
+        ("-0", "14459264155.12895488", Less),
+        ("0", "-14459264155.12895488", Greater),
+        ("-0", "-14459264155.12895488", Greater),
+        // 000 110 same scale
+        ("0.00000000", "38675108055.09052783", Less),
+        ("-0.00000000", "38675108055.09052783", Less),
+        ("0.00000000", "-38675108055.09052783", Greater),
+        ("-0.00000000", "-38675108055.09052783", Greater),
+        // 000 110 different scale
+        ("0.00", "1495767034080324868", Less),
+        ("-0.00", "1495767034080324868", Less),
+        ("0.00", "-1495767034080324868", Greater),
+        ("-0.00", "-1495767034080324868", Greater),
+        // 000 001 same scale
+        ("0.00000000", "359299289270893106016.81305600", Less),
+        ("-0.00000000", "359299289270893106016.81305600", Less),
+        ("0.00000000", "-359299289270893106016.81305600", Greater),
+        ("-0.00000000", "-359299289270893106016.81305600", Greater),
+        // 000 001 different scale
+        ("0.00000000000000000000000000", "261631091689.9518486763536384", Less),
+        ("-0.00000000000000000000000000", "261631091689.9518486763536384", Less),
+        (
+            "0.00000000000000000000000000",
+            "-261631091689.9518486763536384",
+            Greater,
+        ),
+        (
+            "-0.00000000000000000000000000",
+            "-261631091689.9518486763536384",
+            Greater,
+        ),
+        // 000 101 same scale
+        ("0.00000000", "184137107696737410476.63166815", Less),
+        ("-0.00000000", "184137107696737410476.63166815", Less),
+        ("0.00000000", "-184137107696737410476.63166815", Greater),
+        ("-0.00000000", "-184137107696737410476.63166815", Greater),
+        // 000 101 different scale
+        ("0.000000000", "2286857871088.7514840434334478", Less),
+        ("-0.000000000", "2286857871088.7514840434334478", Less),
+        ("0.000000000", "-2286857871088.7514840434334478", Greater),
+        ("-0.000000000", "-2286857871088.7514840434334478", Greater),
+        // 000 011 same scale
+        ("0.00000000", "169194696640288819908.07715840", Less),
+        ("-0.00000000", "169194696640288819908.07715840", Less),
+        ("0.00000000", "-169194696640288819908.07715840", Greater),
+        ("-0.00000000", "-169194696640288819908.07715840", Greater),
+        // 000 011 different scale
+        ("0.00000000", "2757550691.7650076909569048576", Less),
+        ("-0.00000000", "2757550691.7650076909569048576", Less),
+        ("0.00000000", "-2757550691.7650076909569048576", Greater),
+        ("-0.00000000", "-2757550691.7650076909569048576", Greater),
+        // 000 111 same scale
+        ("0.00000000", "133610725292915899001.10059212", Less),
+        ("-0.00000000", "133610725292915899001.10059212", Less),
+        ("0.00000000", "-133610725292915899001.10059212", Greater),
+        ("-0.00000000", "-133610725292915899001.10059212", Greater),
+        // 000 111 different scale
+        ("0.00000000000000000000", "86.25432767926620368822165265", Less),
+        ("-0.00000000000000000000", "86.25432767926620368822165265", Less),
+        ("0.00000000000000000000", "-86.25432767926620368822165265", Greater),
+        ("-0.00000000000000000000", "-86.25432767926620368822165265", Greater),
+        // 100 equality
+        ("0.0000000000598992228", "0.0000000000598992228", Equal),
+        // 100 000 same scale
+        ("0.0000000000598992228", "0.0000000000000000000", Greater),
+        ("-0.0000000000598992228", "0.0000000000000000000", Less),
+        ("0.0000000000598992228", "-0.0000000000000000000", Greater),
+        ("-0.0000000000598992228", "-0.0000000000000000000", Less),
+        // 100 000 different scale
+        ("0.1797407597", "0.0000000000000000000", Greater),
+        ("-0.1797407597", "0.0000000000000000000", Less),
+        ("0.1797407597", "-0.0000000000000000000", Greater),
+        ("-0.1797407597", "-0.0000000000000000000", Less),
+        // 100 100 same scale
+        ("0.0000000000598992228", "0.0000000000064510789", Greater),
+        ("-0.0000000000598992228", "0.0000000000064510789", Less),
+        ("0.0000000000598992228", "-0.0000000000064510789", Greater),
+        ("-0.0000000000598992228", "-0.0000000000064510789", Less),
+        // 100 100 different scale
+        ("0.000000000000011217354", "0.0000000000217735186", Less),
+        ("-0.000000000000011217354", "0.0000000000217735186", Less),
+        ("0.000000000000011217354", "-0.0000000000217735186", Greater),
+        ("-0.000000000000011217354", "-0.0000000000217735186", Greater),
+        // 100 010 same scale
+        ("0.0000000000598992228", "0.0659116848159129600", Less),
+        ("-0.0000000000598992228", "0.0659116848159129600", Less),
+        ("0.0000000000598992228", "-0.0659116848159129600", Greater),
+        ("-0.0000000000598992228", "-0.0659116848159129600", Greater),
+        // 100 010 different scale
+        ("0.00042035421", "0.004709460588143575040", Less),
+        ("-0.00042035421", "0.004709460588143575040", Less),
+        ("0.00042035421", "-0.004709460588143575040", Greater),
+        ("-0.00042035421", "-0.004709460588143575040", Greater),
+        // 100 110 same scale
+        ("0.0000000000598992228", "0.0755686585127091375", Less),
+        ("-0.0000000000598992228", "0.0755686585127091375", Less),
+        ("0.0000000000598992228", "-0.0755686585127091375", Greater),
+        ("-0.0000000000598992228", "-0.0755686585127091375", Greater),
+        // 100 110 different scale
+        ("14872.94465", "1284707831905.854085", Less),
+        ("-14872.94465", "1284707831905.854085", Less),
+        ("14872.94465", "-1284707831905.854085", Greater),
+        ("-14872.94465", "-1284707831905.854085", Greater),
+        // 100 001 same scale
+        ("0.0000000000598992228", "888767595.6145376468836286464", Less),
+        ("-0.0000000000598992228", "888767595.6145376468836286464", Less),
+        ("0.0000000000598992228", "-888767595.6145376468836286464", Greater),
+        ("-0.0000000000598992228", "-888767595.6145376468836286464", Greater),
+        // 100 001 different scale
+        ("0.0002108155975", "36.07555527243968476014968832", Less),
+        ("-0.0002108155975", "36.07555527243968476014968832", Less),
+        ("0.0002108155975", "-36.07555527243968476014968832", Greater),
+        ("-0.0002108155975", "-36.07555527243968476014968832", Greater),
+        // 100 101 same scale
+        ("0.0000000000598992228", "125730345.7412344676309569911", Less),
+        ("-0.0000000000598992228", "125730345.7412344676309569911", Less),
+        ("0.0000000000598992228", "-125730345.7412344676309569911", Greater),
+        ("-0.0000000000598992228", "-125730345.7412344676309569911", Greater),
+        // 100 101 different scale
+        ("0.0000000000871576741", "1.7283925558865766140690239662", Less),
+        ("-0.0000000000871576741", "1.7283925558865766140690239662", Less),
+        ("0.0000000000871576741", "-1.7283925558865766140690239662", Greater),
+        ("-0.0000000000871576741", "-1.7283925558865766140690239662", Greater),
+        // 100 011 same scale
+        ("0.0000000000598992228", "645513262.9193254090737451008", Less),
+        ("-0.0000000000598992228", "645513262.9193254090737451008", Less),
+        ("0.0000000000598992228", "-645513262.9193254090737451008", Greater),
+        ("-0.0000000000598992228", "-645513262.9193254090737451008", Greater),
+        // 100 011 different scale
+        ("0.000000000000000760885021", "3718370638.2004059326675681280", Less),
+        ("-0.000000000000000760885021", "3718370638.2004059326675681280", Less),
+        ("0.000000000000000760885021", "-3718370638.2004059326675681280", Greater),
+        (
+            "-0.000000000000000760885021",
+            "-3718370638.2004059326675681280",
+            Greater,
+        ),
+        // 100 111 same scale
+        ("0.0000000000598992228", "422482675.5515775479939306437", Less),
+        ("-0.0000000000598992228", "422482675.5515775479939306437", Less),
+        ("0.0000000000598992228", "-422482675.5515775479939306437", Greater),
+        ("-0.0000000000598992228", "-422482675.5515775479939306437", Greater),
+        // 100 111 different scale
+        ("0.000000044182898", "25.452953262109919998605674045", Less),
+        ("-0.000000044182898", "25.452953262109919998605674045", Less),
+        ("0.000000044182898", "-25.452953262109919998605674045", Greater),
+        ("-0.000000044182898", "-25.452953262109919998605674045", Greater),
+        // 010 equality
+        ("423.1744746042687488", "423.1744746042687488", Equal),
+        // 010 000 same scale
+        ("423.1744746042687488", "0.0000000000000000", Greater),
+        ("-423.1744746042687488", "0.0000000000000000", Less),
+        ("423.1744746042687488", "-0.0000000000000000", Greater),
+        ("-423.1744746042687488", "-0.0000000000000000", Less),
+        // 010 000 different scale
+        ("9002354991.192604672", "0.00000000000000000000000000", Greater),
+        ("-9002354991.192604672", "0.00000000000000000000000000", Less),
+        ("9002354991.192604672", "-0.00000000000000000000000000", Greater),
+        ("-9002354991.192604672", "-0.00000000000000000000000000", Less),
+        // 010 100 same scale
+        ("423.1744746042687488", "0.0000000981820809", Greater),
+        ("-423.1744746042687488", "0.0000000981820809", Less),
+        ("423.1744746042687488", "-0.0000000981820809", Greater),
+        ("-423.1744746042687488", "-0.0000000981820809", Less),
+        // 010 100 different scale
+        ("4327019125101559.808", "0.00484846050", Greater),
+        ("-4327019125101559.808", "0.00484846050", Less),
+        ("4327019125101559.808", "-0.00484846050", Greater),
+        ("-4327019125101559.808", "-0.00484846050", Less),
+        // 010 010 same scale
+        ("423.1744746042687488", "786.9082854590775296", Less),
+        ("-423.1744746042687488", "786.9082854590775296", Less),
+        ("423.1744746042687488", "-786.9082854590775296", Greater),
+        ("-423.1744746042687488", "-786.9082854590775296", Greater),
+        // 010 010 different scale
+        ("793.0067125291450368", "0.0001587297248335626240", Greater),
+        ("-793.0067125291450368", "0.0001587297248335626240", Less),
+        ("793.0067125291450368", "-0.0001587297248335626240", Greater),
+        ("-793.0067125291450368", "-0.0001587297248335626240", Less),
+        // 010 110 same scale
+        ("423.1744746042687488", "300.0541360230572049", Greater),
+        ("-423.1744746042687488", "300.0541360230572049", Less),
+        ("423.1744746042687488", "-300.0541360230572049", Greater),
+        ("-423.1744746042687488", "-300.0541360230572049", Less),
+        // 010 110 different scale
+        ("90627.2042582540288", "4472414566654924.741", Less),
+        ("-90627.2042582540288", "4472414566654924.741", Less),
+        ("90627.2042582540288", "-4472414566654924.741", Greater),
+        ("-90627.2042582540288", "-4472414566654924.741", Greater),
+        // 010 001 same scale
+        ("423.1744746042687488", "3960577151543.5796707636412416", Less),
+        ("-423.1744746042687488", "3960577151543.5796707636412416", Less),
+        ("423.1744746042687488", "-3960577151543.5796707636412416", Greater),
+        ("-423.1744746042687488", "-3960577151543.5796707636412416", Greater),
+        // 010 001 different scale
+        ("0.000008867461591822499840", "185286.04378228713249986052096", Less),
+        ("-0.000008867461591822499840", "185286.04378228713249986052096", Less),
+        ("0.000008867461591822499840", "-185286.04378228713249986052096", Greater),
+        (
+            "-0.000008867461591822499840",
+            "-185286.04378228713249986052096",
+            Greater,
+        ),
+        // 010 101 same scale
+        ("423.1744746042687488", "2825958416017.6213507229869501", Less),
+        ("-423.1744746042687488", "2825958416017.6213507229869501", Less),
+        ("423.1744746042687488", "-2825958416017.6213507229869501", Greater),
+        ("-423.1744746042687488", "-2825958416017.6213507229869501", Greater),
+        // 010 101 different scale
+        ("0.01901870767742648320", "37662082383021542232.529651212", Less),
+        ("-0.01901870767742648320", "37662082383021542232.529651212", Less),
+        ("0.01901870767742648320", "-37662082383021542232.529651212", Greater),
+        ("-0.01901870767742648320", "-37662082383021542232.529651212", Greater),
+        // 010 011 same scale
+        ("423.1744746042687488", "3628063966991.6759417059016704", Less),
+        ("-423.1744746042687488", "3628063966991.6759417059016704", Less),
+        ("423.1744746042687488", "-3628063966991.6759417059016704", Greater),
+        ("-423.1744746042687488", "-3628063966991.6759417059016704", Greater),
+        // 010 011 different scale
+        ("45359904.32470925312", "2.4203452488052342918570049536", Greater),
+        ("-45359904.32470925312", "2.4203452488052342918570049536", Less),
+        ("45359904.32470925312", "-2.4203452488052342918570049536", Greater),
+        ("-45359904.32470925312", "-2.4203452488052342918570049536", Less),
+        // 010 111 same scale
+        ("423.1744746042687488", "2629665172331.9610693820109120", Less),
+        ("-423.1744746042687488", "2629665172331.9610693820109120", Less),
+        ("423.1744746042687488", "-2629665172331.9610693820109120", Greater),
+        ("-423.1744746042687488", "-2629665172331.9610693820109120", Greater),
+        // 010 111 different scale
+        ("0.0006420803252266205184", "8172.032417576265900588945489", Less),
+        ("-0.0006420803252266205184", "8172.032417576265900588945489", Less),
+        ("0.0006420803252266205184", "-8172.032417576265900588945489", Greater),
+        ("-0.0006420803252266205184", "-8172.032417576265900588945489", Greater),
+        // 110 equality
+        ("844.5530620517286511", "844.5530620517286511", Equal),
+        // 110 000 same scale
+        ("844.5530620517286511", "0.0000000000000000", Greater),
+        ("-844.5530620517286511", "0.0000000000000000", Less),
+        ("844.5530620517286511", "-0.0000000000000000", Greater),
+        ("-844.5530620517286511", "-0.0000000000000000", Less),
+        // 110 000 different scale
+        ("3285.530033386074797", "0.00000000000000000000", Greater),
+        ("-3285.530033386074797", "0.00000000000000000000", Less),
+        ("3285.530033386074797", "-0.00000000000000000000", Greater),
+        ("-3285.530033386074797", "-0.00000000000000000000", Less),
+        // 110 100 same scale
+        ("844.5530620517286511", "0.0000001953470063", Greater),
+        ("-844.5530620517286511", "0.0000001953470063", Less),
+        ("844.5530620517286511", "-0.0000001953470063", Greater),
+        ("-844.5530620517286511", "-0.0000001953470063", Less),
+        // 110 100 different scale
+        ("371284.0210972493371", "0.000000000000001307794657", Greater),
+        ("-371284.0210972493371", "0.000000000000001307794657", Less),
+        ("371284.0210972493371", "-0.000000000000001307794657", Greater),
+        ("-371284.0210972493371", "-0.000000000000001307794657", Less),
+        // 110 010 same scale
+        ("844.5530620517286511", "612.1542773033140224", Greater),
+        ("-844.5530620517286511", "612.1542773033140224", Less),
+        ("844.5530620517286511", "-612.1542773033140224", Greater),
+        ("-844.5530620517286511", "-612.1542773033140224", Less),
+        // 110 010 different scale
+        ("0.00004869219159821525572", "23341676485159.15776", Less),
+        ("-0.00004869219159821525572", "23341676485159.15776", Less),
+        ("0.00004869219159821525572", "-23341676485159.15776", Greater),
+        ("-0.00004869219159821525572", "-23341676485159.15776", Greater),
+        // 110 110 same scale
+        ("844.5530620517286511", "326.6132818317622015", Greater),
+        ("-844.5530620517286511", "326.6132818317622015", Less),
+        ("844.5530620517286511", "-326.6132818317622015", Greater),
+        ("-844.5530620517286511", "-326.6132818317622015", Less),
+        // 110 110 different scale
+        ("4752139369958.820619", "1330851022882027.972", Less),
+        ("-4752139369958.820619", "1330851022882027.972", Less),
+        ("4752139369958.820619", "-1330851022882027.972", Greater),
+        ("-4752139369958.820619", "-1330851022882027.972", Greater),
+        // 110 001 same scale
+        ("844.5530620517286511", "3585610241942.1922435648192512", Less),
+        ("-844.5530620517286511", "3585610241942.1922435648192512", Less),
+        ("844.5530620517286511", "-3585610241942.1922435648192512", Greater),
+        ("-844.5530620517286511", "-3585610241942.1922435648192512", Greater),
+        // 110 001 different scale
+        ("539313715923.4424678", "12410950080603997079634706432", Less),
+        ("-539313715923.4424678", "12410950080603997079634706432", Less),
+        ("539313715923.4424678", "-12410950080603997079634706432", Greater),
+        ("-539313715923.4424678", "-12410950080603997079634706432", Greater),
+        // 110 101 same scale
+        ("844.5530620517286511", "1947825396031.6933708908343230", Less),
+        ("-844.5530620517286511", "1947825396031.6933708908343230", Less),
+        ("844.5530620517286511", "-1947825396031.6933708908343230", Greater),
+        ("-844.5530620517286511", "-1947825396031.6933708908343230", Greater),
+        // 110 101 different scale
+        ("0.2301405445512525317", "245433629587.71206426704897154", Less),
+        ("-0.2301405445512525317", "245433629587.71206426704897154", Less),
+        ("0.2301405445512525317", "-245433629587.71206426704897154", Greater),
+        ("-0.2301405445512525317", "-245433629587.71206426704897154", Greater),
+        // 110 011 same scale
+        ("844.5530620517286511", "3637850451015.0843464291450880", Less),
+        ("-844.5530620517286511", "3637850451015.0843464291450880", Less),
+        ("844.5530620517286511", "-3637850451015.0843464291450880", Greater),
+        ("-844.5530620517286511", "-3637850451015.0843464291450880", Greater),
+        // 110 011 different scale
+        ("0.00000000717944802566514691", "18143443615.480512395717115904", Less),
+        ("-0.00000000717944802566514691", "18143443615.480512395717115904", Less),
+        (
+            "0.00000000717944802566514691",
+            "-18143443615.480512395717115904",
+            Greater,
+        ),
+        (
+            "-0.00000000717944802566514691",
+            "-18143443615.480512395717115904",
+            Greater,
+        ),
+        // 110 111 same scale
+        ("844.5530620517286511", "2738424264600.4917875777303163", Less),
+        ("-844.5530620517286511", "2738424264600.4917875777303163", Less),
+        ("844.5530620517286511", "-2738424264600.4917875777303163", Greater),
+        ("-844.5530620517286511", "-2738424264600.4917875777303163", Greater),
+        // 110 111 different scale
+        ("0.0000000007762706076409491335", "2489879185787497651.6458518595", Less),
+        (
+            "-0.0000000007762706076409491335",
+            "2489879185787497651.6458518595",
+            Less,
+        ),
+        (
+            "0.0000000007762706076409491335",
+            "-2489879185787497651.6458518595",
+            Greater,
+        ),
+        (
+            "-0.0000000007762706076409491335",
+            "-2489879185787497651.6458518595",
+            Greater,
+        ),
+        // 001 equality
+        ("316007568232.9263258873102336", "316007568232.9263258873102336", Equal),
+        // 001 000 same scale
+        ("316007568232.9263258873102336", "0.0000000000000000", Greater),
+        ("-316007568232.9263258873102336", "0.0000000000000000", Less),
+        ("316007568232.9263258873102336", "-0.0000000000000000", Greater),
+        ("-316007568232.9263258873102336", "-0.0000000000000000", Less),
+        // 001 000 different scale
+        ("3522055990024364385815547084.8", "0.000000000000", Greater),
+        ("-3522055990024364385815547084.8", "0.000000000000", Less),
+        ("3522055990024364385815547084.8", "-0.000000000000", Greater),
+        ("-3522055990024364385815547084.8", "-0.000000000000", Less),
+        // 001 100 same scale
+        ("316007568232.9263258873102336", "0.0000001073412971", Greater),
+        ("-316007568232.9263258873102336", "0.0000001073412971", Less),
+        ("316007568232.9263258873102336", "-0.0000001073412971", Greater),
+        ("-316007568232.9263258873102336", "-0.0000001073412971", Less),
+        // 001 100 different scale
+        ("1319006.0491408208640440532992", "0.00000000000000611866432", Greater),
+        ("-1319006.0491408208640440532992", "0.00000000000000611866432", Less),
+        ("1319006.0491408208640440532992", "-0.00000000000000611866432", Greater),
+        ("-1319006.0491408208640440532992", "-0.00000000000000611866432", Less),
+        // 001 010 same scale
+        ("316007568232.9263258873102336", "159.1054215143227392", Greater),
+        ("-316007568232.9263258873102336", "159.1054215143227392", Less),
+        ("316007568232.9263258873102336", "-159.1054215143227392", Greater),
+        ("-316007568232.9263258873102336", "-159.1054215143227392", Less),
+        // 001 010 different scale
+        ("2470144.7146711063666704252928", "211.0186916505714688", Greater),
+        ("-2470144.7146711063666704252928", "211.0186916505714688", Less),
+        ("2470144.7146711063666704252928", "-211.0186916505714688", Greater),
+        ("-2470144.7146711063666704252928", "-211.0186916505714688", Less),
+        // 001 110 same scale
+        ("316007568232.9263258873102336", "15.1186658969096112", Greater),
+        ("-316007568232.9263258873102336", "15.1186658969096112", Less),
+        ("316007568232.9263258873102336", "-15.1186658969096112", Greater),
+        ("-316007568232.9263258873102336", "-15.1186658969096112", Less),
+        // 001 110 different scale
+        ("3840504199004148630832.3360768", "7581138850996748864", Greater),
+        ("-3840504199004148630832.3360768", "7581138850996748864", Less),
+        ("3840504199004148630832.3360768", "-7581138850996748864", Greater),
+        ("-3840504199004148630832.3360768", "-7581138850996748864", Less),
+        // 001 001 same scale
+        ("316007568232.9263258873102336", "810157633226.6053390856880128", Less),
+        ("-316007568232.9263258873102336", "810157633226.6053390856880128", Less),
+        (
+            "316007568232.9263258873102336",
+            "-810157633226.6053390856880128",
+            Greater,
+        ),
+        (
+            "-316007568232.9263258873102336",
+            "-810157633226.6053390856880128",
+            Greater,
+        ),
+        // 001 001 different scale
+        ("1951046382014.4037956952260608", "3626102772868740412010083.1232", Less),
+        (
+            "-1951046382014.4037956952260608",
+            "3626102772868740412010083.1232",
+            Less,
+        ),
+        (
+            "1951046382014.4037956952260608",
+            "-3626102772868740412010083.1232",
+            Greater,
+        ),
+        (
+            "-1951046382014.4037956952260608",
+            "-3626102772868740412010083.1232",
+            Greater,
+        ),
+        // 001 101 same scale
+        ("316007568232.9263258873102336", "3258394380359.1965879291312453", Less),
+        ("-316007568232.9263258873102336", "3258394380359.1965879291312453", Less),
+        (
+            "316007568232.9263258873102336",
+            "-3258394380359.1965879291312453",
+            Greater,
+        ),
+        (
+            "-316007568232.9263258873102336",
+            "-3258394380359.1965879291312453",
+            Greater,
+        ),
+        // 001 101 different scale
+        (
+            "17580513970289834.943527780352",
+            "3.7977957031395371036126086595",
+            Greater,
+        ),
+        (
+            "-17580513970289834.943527780352",
+            "3.7977957031395371036126086595",
+            Less,
+        ),
+        (
+            "17580513970289834.943527780352",
+            "-3.7977957031395371036126086595",
+            Greater,
+        ),
+        (
+            "-17580513970289834.943527780352",
+            "-3.7977957031395371036126086595",
+            Less,
+        ),
+        // 001 011 same scale
+        ("316007568232.9263258873102336", "1154574080460.9867510617997312", Less),
+        ("-316007568232.9263258873102336", "1154574080460.9867510617997312", Less),
+        (
+            "316007568232.9263258873102336",
+            "-1154574080460.9867510617997312",
+            Greater,
+        ),
+        (
+            "-316007568232.9263258873102336",
+            "-1154574080460.9867510617997312",
+            Greater,
+        ),
+        // 001 011 different scale
+        ("2008379587.5525351789031325696", "32824109460554.341800487157760", Less),
+        (
+            "-2008379587.5525351789031325696",
+            "32824109460554.341800487157760",
+            Less,
+        ),
+        (
+            "2008379587.5525351789031325696",
+            "-32824109460554.341800487157760",
+            Greater,
+        ),
+        (
+            "-2008379587.5525351789031325696",
+            "-32824109460554.341800487157760",
+            Greater,
+        ),
+        // 001 111 same scale
+        ("316007568232.9263258873102336", "2816795479724.6069787794805311", Less),
+        ("-316007568232.9263258873102336", "2816795479724.6069787794805311", Less),
+        (
+            "316007568232.9263258873102336",
+            "-2816795479724.6069787794805311",
+            Greater,
+        ),
+        (
+            "-316007568232.9263258873102336",
+            "-2816795479724.6069787794805311",
+            Greater,
+        ),
+        // 001 111 different scale
+        (
+            "3536806574745420013541890.4576",
+            "9793146.81730411145833529126",
+            Greater,
+        ),
+        ("-3536806574745420013541890.4576", "9793146.81730411145833529126", Less),
+        (
+            "3536806574745420013541890.4576",
+            "-9793146.81730411145833529126",
+            Greater,
+        ),
+        ("-3536806574745420013541890.4576", "-9793146.81730411145833529126", Less),
+        // 101 equality
+        (
+            "254208186622762823842.71629992",
+            "254208186622762823842.71629992",
+            Equal,
+        ),
+        // 101 000 same scale
+        ("254208186622762823842.71629992", "0.00000000", Greater),
+        ("-254208186622762823842.71629992", "0.00000000", Less),
+        ("254208186622762823842.71629992", "-0.00000000", Greater),
+        ("-254208186622762823842.71629992", "-0.00000000", Less),
+        // 101 000 different scale
+        ("975421950664039.3614304091804", "0.0000000000000000", Greater),
+        ("-975421950664039.3614304091804", "0.0000000000000000", Less),
+        ("975421950664039.3614304091804", "-0.0000000000000000", Greater),
+        ("-975421950664039.3614304091804", "-0.0000000000000000", Less),
+        // 101 100 same scale
+        ("254208186622762823842.71629992", "10.74141379", Greater),
+        ("-254208186622762823842.71629992", "10.74141379", Less),
+        ("254208186622762823842.71629992", "-10.74141379", Greater),
+        ("-254208186622762823842.71629992", "-10.74141379", Less),
+        // 101 100 different scale
+        ("2552221405032275.8358630506229", "0.000000000000000592656995", Greater),
+        ("-2552221405032275.8358630506229", "0.000000000000000592656995", Less),
+        ("2552221405032275.8358630506229", "-0.000000000000000592656995", Greater),
+        ("-2552221405032275.8358630506229", "-0.000000000000000592656995", Less),
+        // 101 010 same scale
+        ("254208186622762823842.71629992", "62767493748.48499712", Greater),
+        ("-254208186622762823842.71629992", "62767493748.48499712", Less),
+        ("254208186622762823842.71629992", "-62767493748.48499712", Greater),
+        ("-254208186622762823842.71629992", "-62767493748.48499712", Less),
+        // 101 010 different scale
+        ("197346074219.25327589999174264", "0.7623493575178715136", Greater),
+        ("-197346074219.25327589999174264", "0.7623493575178715136", Less),
+        ("197346074219.25327589999174264", "-0.7623493575178715136", Greater),
+        ("-197346074219.25327589999174264", "-0.7623493575178715136", Less),
+        // 101 110 same scale
+        ("254208186622762823842.71629992", "76597126194.51389094", Greater),
+        ("-254208186622762823842.71629992", "76597126194.51389094", Less),
+        ("254208186622762823842.71629992", "-76597126194.51389094", Greater),
+        ("-254208186622762823842.71629992", "-76597126194.51389094", Less),
+        // 101 110 different scale
+        ("25899773651648.380071130043467", "61306258142804903.38", Less),
+        ("-25899773651648.380071130043467", "61306258142804903.38", Less),
+        ("25899773651648.380071130043467", "-61306258142804903.38", Greater),
+        ("-25899773651648.380071130043467", "-61306258142804903.38", Greater),
+        // 101 001 same scale
+        (
+            "254208186622762823842.71629992",
+            "83547191565151621967.28086528",
+            Greater,
+        ),
+        ("-254208186622762823842.71629992", "83547191565151621967.28086528", Less),
+        (
+            "254208186622762823842.71629992",
+            "-83547191565151621967.28086528",
+            Greater,
+        ),
+        (
+            "-254208186622762823842.71629992",
+            "-83547191565151621967.28086528",
+            Less,
+        ),
+        // 101 001 different scale
+        ("244762.90302171293318719286219", "9964072.600255221193011888128", Less),
+        ("-244762.90302171293318719286219", "9964072.600255221193011888128", Less),
+        (
+            "244762.90302171293318719286219",
+            "-9964072.600255221193011888128",
+            Greater,
+        ),
+        (
+            "-244762.90302171293318719286219",
+            "-9964072.600255221193011888128",
+            Greater,
+        ),
+        // 101 101 same scale
+        (
+            "254208186622762823842.71629992",
+            "106541875981662806348.63716235",
+            Greater,
+        ),
+        (
+            "-254208186622762823842.71629992",
+            "106541875981662806348.63716235",
+            Less,
+        ),
+        (
+            "254208186622762823842.71629992",
+            "-106541875981662806348.63716235",
+            Greater,
+        ),
+        (
+            "-254208186622762823842.71629992",
+            "-106541875981662806348.63716235",
+            Less,
+        ),
+        // 101 101 different scale
+        ("362319.18250030256385507568342", "3619454249020577423546109236", Less),
+        ("-362319.18250030256385507568342", "3619454249020577423546109236", Less),
+        (
+            "362319.18250030256385507568342",
+            "-3619454249020577423546109236",
+            Greater,
+        ),
+        (
+            "-362319.18250030256385507568342",
+            "-3619454249020577423546109236",
+            Greater,
+        ),
+        // 101 011 same scale
+        (
+            "254208186622762823842.71629992",
+            "156781478378762688050.06557184",
+            Greater,
+        ),
+        (
+            "-254208186622762823842.71629992",
+            "156781478378762688050.06557184",
+            Less,
+        ),
+        (
+            "254208186622762823842.71629992",
+            "-156781478378762688050.06557184",
+            Greater,
+        ),
+        (
+            "-254208186622762823842.71629992",
+            "-156781478378762688050.06557184",
+            Less,
+        ),
+        // 101 011 different scale
+        ("2486073465266.0337130589931876", "153874906950888902858.19691008", Less),
+        (
+            "-2486073465266.0337130589931876",
+            "153874906950888902858.19691008",
+            Less,
+        ),
+        (
+            "2486073465266.0337130589931876",
+            "-153874906950888902858.19691008",
+            Greater,
+        ),
+        (
+            "-2486073465266.0337130589931876",
+            "-153874906950888902858.19691008",
+            Greater,
+        ),
+        // 101 111 same scale
+        (
+            "254208186622762823842.71629992",
+            "10101645723744656462.08148676",
+            Greater,
+        ),
+        ("-254208186622762823842.71629992", "10101645723744656462.08148676", Less),
+        (
+            "254208186622762823842.71629992",
+            "-10101645723744656462.08148676",
+            Greater,
+        ),
+        (
+            "-254208186622762823842.71629992",
+            "-10101645723744656462.08148676",
+            Less,
+        ),
+        // 101 111 different scale
+        ("14107601965.909635434653634526", "111238758546502973973477.99454", Less),
+        (
+            "-14107601965.909635434653634526",
+            "111238758546502973973477.99454",
+            Less,
+        ),
+        (
+            "14107601965.909635434653634526",
+            "-111238758546502973973477.99454",
+            Greater,
+        ),
+        (
+            "-14107601965.909635434653634526",
+            "-111238758546502973973477.99454",
+            Greater,
+        ),
+        // 011 equality
+        (
+            "272322.48219624218537039495168",
+            "272322.48219624218537039495168",
+            Equal,
+        ),
+        // 011 000 same scale
+        ("272322.48219624218537039495168", "0.00000000000000000000000", Greater),
+        ("-272322.48219624218537039495168", "0.00000000000000000000000", Less),
+        ("272322.48219624218537039495168", "-0.00000000000000000000000", Greater),
+        ("-272322.48219624218537039495168", "-0.00000000000000000000000", Less),
+        // 011 000 different scale
+        ("3214885516.0787854158246969344", "0.00000000000000000", Greater),
+        ("-3214885516.0787854158246969344", "0.00000000000000000", Less),
+        ("3214885516.0787854158246969344", "-0.00000000000000000", Greater),
+        ("-3214885516.0787854158246969344", "-0.00000000000000000", Less),
+        // 011 100 same scale
+        ("272322.48219624218537039495168", "0.00000000000000379487994", Greater),
+        ("-272322.48219624218537039495168", "0.00000000000000379487994", Less),
+        ("272322.48219624218537039495168", "-0.00000000000000379487994", Greater),
+        ("-272322.48219624218537039495168", "-0.00000000000000379487994", Less),
+        // 011 100 different scale
+        ("388166715906.19371912596291584", "0.000000000000001996700736", Greater),
+        ("-388166715906.19371912596291584", "0.000000000000001996700736", Less),
+        ("388166715906.19371912596291584", "-0.000000000000001996700736", Greater),
+        ("-388166715906.19371912596291584", "-0.000000000000001996700736", Less),
+        // 011 010 same scale
+        ("272322.48219624218537039495168", "0.00000175873997328613376", Greater),
+        ("-272322.48219624218537039495168", "0.00000175873997328613376", Less),
+        ("272322.48219624218537039495168", "-0.00000175873997328613376", Greater),
+        ("-272322.48219624218537039495168", "-0.00000175873997328613376", Less),
+        // 011 010 different scale
+        (
+            "17963864.0946434527121637376",
+            "0.0000000001112699367808040960",
+            Greater,
+        ),
+        ("-17963864.0946434527121637376", "0.0000000001112699367808040960", Less),
+        (
+            "17963864.0946434527121637376",
+            "-0.0000000001112699367808040960",
+            Greater,
+        ),
+        ("-17963864.0946434527121637376", "-0.0000000001112699367808040960", Less),
+        // 011 110 same scale
+        ("272322.48219624218537039495168", "0.00009168252278596474115", Greater),
+        ("-272322.48219624218537039495168", "0.00009168252278596474115", Less),
+        ("272322.48219624218537039495168", "-0.00009168252278596474115", Greater),
+        ("-272322.48219624218537039495168", "-0.00009168252278596474115", Less),
+        // 011 110 different scale
+        ("3.1530040332824324172729548800", "0.05636139104712652784", Greater),
+        ("-3.1530040332824324172729548800", "0.05636139104712652784", Less),
+        ("3.1530040332824324172729548800", "-0.05636139104712652784", Greater),
+        ("-3.1530040332824324172729548800", "-0.05636139104712652784", Less),
+        // 011 001 same scale
+        (
+            "272322.48219624218537039495168",
+            "78956.15667671475190631497728",
+            Greater,
+        ),
+        ("-272322.48219624218537039495168", "78956.15667671475190631497728", Less),
+        (
+            "272322.48219624218537039495168",
+            "-78956.15667671475190631497728",
+            Greater,
+        ),
+        (
+            "-272322.48219624218537039495168",
+            "-78956.15667671475190631497728",
+            Less,
+        ),
+        // 011 001 different scale
+        (
+            "247159638929.90774677497446400",
+            "18573048.37697991870462820352",
+            Greater,
+        ),
+        ("-247159638929.90774677497446400", "18573048.37697991870462820352", Less),
+        (
+            "247159638929.90774677497446400",
+            "-18573048.37697991870462820352",
+            Greater,
+        ),
+        (
+            "-247159638929.90774677497446400",
+            "-18573048.37697991870462820352",
+            Less,
+        ),
+        // 011 101 same scale
+        ("272322.48219624218537039495168", "311953.28803357654172947915942", Less),
+        (
+            "-272322.48219624218537039495168",
+            "311953.28803357654172947915942",
+            Less,
+        ),
+        (
+            "272322.48219624218537039495168",
+            "-311953.28803357654172947915942",
+            Greater,
+        ),
+        (
+            "-272322.48219624218537039495168",
+            "-311953.28803357654172947915942",
+            Greater,
+        ),
+        // 011 101 different scale
+        ("2922696937234119470.0273745920", "22441101906503785827606686629", Less),
+        ("-2922696937234119470.0273745920", "22441101906503785827606686629", Less),
+        (
+            "2922696937234119470.0273745920",
+            "-22441101906503785827606686629",
+            Greater,
+        ),
+        (
+            "-2922696937234119470.0273745920",
+            "-22441101906503785827606686629",
+            Greater,
+        ),
+        // 011 011 same scale
+        ("272322.48219624218537039495168", "348316.28306497394164006649856", Less),
+        (
+            "-272322.48219624218537039495168",
+            "348316.28306497394164006649856",
+            Less,
+        ),
+        (
+            "272322.48219624218537039495168",
+            "-348316.28306497394164006649856",
+            Greater,
+        ),
+        (
+            "-272322.48219624218537039495168",
+            "-348316.28306497394164006649856",
+            Greater,
+        ),
+        // 011 011 different scale
+        ("178190346.76624086261395619840", "62208030746.22038852927225856", Less),
+        ("-178190346.76624086261395619840", "62208030746.22038852927225856", Less),
+        (
+            "178190346.76624086261395619840",
+            "-62208030746.22038852927225856",
+            Greater,
+        ),
+        (
+            "-178190346.76624086261395619840",
+            "-62208030746.22038852927225856",
+            Greater,
+        ),
+        // 011 111 same scale
+        (
+            "272322.48219624218537039495168",
+            "41534.52021391898039335355927",
+            Greater,
+        ),
+        ("-272322.48219624218537039495168", "41534.52021391898039335355927", Less),
+        (
+            "272322.48219624218537039495168",
+            "-41534.52021391898039335355927",
+            Greater,
+        ),
+        (
+            "-272322.48219624218537039495168",
+            "-41534.52021391898039335355927",
+            Less,
+        ),
+        // 011 111 different scale
+        ("11.959910518519677083499626496", "2844684364802261541879551.2259", Less),
+        (
+            "-11.959910518519677083499626496",
+            "2844684364802261541879551.2259",
+            Less,
+        ),
+        (
+            "11.959910518519677083499626496",
+            "-2844684364802261541879551.2259",
+            Greater,
+        ),
+        (
+            "-11.959910518519677083499626496",
+            "-2844684364802261541879551.2259",
+            Greater,
+        ),
+        // 111 equality
+        (
+            "3836286746260530032892706.6174",
+            "3836286746260530032892706.6174",
+            Equal,
+        ),
+        // 111 000 same scale
+        ("3836286746260530032892706.6174", "0.0000", Greater),
+        ("-3836286746260530032892706.6174", "0.0000", Less),
+        ("3836286746260530032892706.6174", "-0.0000", Greater),
+        ("-3836286746260530032892706.6174", "-0.0000", Less),
+        // 111 000 different scale
+        ("4401861854803552.033657814547", "0.0000000", Greater),
+        ("-4401861854803552.033657814547", "0.0000000", Less),
+        ("4401861854803552.033657814547", "-0.0000000", Greater),
+        ("-4401861854803552.033657814547", "-0.0000000", Less),
+        // 111 100 same scale
+        ("3836286746260530032892706.6174", "68758.6561", Greater),
+        ("-3836286746260530032892706.6174", "68758.6561", Less),
+        ("3836286746260530032892706.6174", "-68758.6561", Greater),
+        ("-3836286746260530032892706.6174", "-68758.6561", Less),
+        // 111 100 different scale
+        (
+            "18794337354296131695536777.153",
+            "0.000000000000000001563875977",
+            Greater,
+        ),
+        ("-18794337354296131695536777.153", "0.000000000000000001563875977", Less),
+        (
+            "18794337354296131695536777.153",
+            "-0.000000000000000001563875977",
+            Greater,
+        ),
+        (
+            "-18794337354296131695536777.153",
+            "-0.000000000000000001563875977",
+            Less,
+        ),
+        // 111 010 same scale
+        ("3836286746260530032892706.6174", "439097665563236.7616", Greater),
+        ("-3836286746260530032892706.6174", "439097665563236.7616", Less),
+        ("3836286746260530032892706.6174", "-439097665563236.7616", Greater),
+        ("-3836286746260530032892706.6174", "-439097665563236.7616", Less),
+        // 111 010 different scale
+        ("219364497389.57405761662363679", "0.6569644274462228480", Greater),
+        ("-219364497389.57405761662363679", "0.6569644274462228480", Less),
+        ("219364497389.57405761662363679", "-0.6569644274462228480", Greater),
+        ("-219364497389.57405761662363679", "-0.6569644274462228480", Less),
+        // 111 110 same scale
+        ("3836286746260530032892706.6174", "100013274294974.8269", Greater),
+        ("-3836286746260530032892706.6174", "100013274294974.8269", Less),
+        ("3836286746260530032892706.6174", "-100013274294974.8269", Greater),
+        ("-3836286746260530032892706.6174", "-100013274294974.8269", Less),
+        // 111 110 different scale
+        ("76072704083682.85472479207171", "50.52437989651117182", Greater),
+        ("-76072704083682.85472479207171", "50.52437989651117182", Less),
+        ("76072704083682.85472479207171", "-50.52437989651117182", Greater),
+        ("-76072704083682.85472479207171", "-50.52437989651117182", Less),
+        // 111 001 same scale
+        (
+            "3836286746260530032892706.6174",
+            "2766133872545894272402142.0032",
+            Greater,
+        ),
+        (
+            "-3836286746260530032892706.6174",
+            "2766133872545894272402142.0032",
+            Less,
+        ),
+        (
+            "3836286746260530032892706.6174",
+            "-2766133872545894272402142.0032",
+            Greater,
+        ),
+        (
+            "-3836286746260530032892706.6174",
+            "-2766133872545894272402142.0032",
+            Less,
+        ),
+        // 111 001 different scale
+        (
+            "38199979438250010.80610984395",
+            "31104752430710.408848162684928",
+            Greater,
+        ),
+        ("-38199979438250010.80610984395", "31104752430710.408848162684928", Less),
+        (
+            "38199979438250010.80610984395",
+            "-31104752430710.408848162684928",
+            Greater,
+        ),
+        (
+            "-38199979438250010.80610984395",
+            "-31104752430710.408848162684928",
+            Less,
+        ),
+        // 111 101 same scale
+        (
+            "3836286746260530032892706.6174",
+            "441847458119168110406908.6115",
+            Greater,
+        ),
+        ("-3836286746260530032892706.6174", "441847458119168110406908.6115", Less),
+        (
+            "3836286746260530032892706.6174",
+            "-441847458119168110406908.6115",
+            Greater,
+        ),
+        (
+            "-3836286746260530032892706.6174",
+            "-441847458119168110406908.6115",
+            Less,
+        ),
+        // 111 101 different scale
+        ("255945012905633524.15746865235", "1005021647855597114428453997.8", Less),
+        (
+            "-255945012905633524.15746865235",
+            "1005021647855597114428453997.8",
+            Less,
+        ),
+        (
+            "255945012905633524.15746865235",
+            "-1005021647855597114428453997.8",
+            Greater,
+        ),
+        (
+            "-255945012905633524.15746865235",
+            "-1005021647855597114428453997.8",
+            Greater,
+        ),
+        // 111 011 same scale
+        (
+            "3836286746260530032892706.6174",
+            "1111481055212557787730018.3040",
+            Greater,
+        ),
+        (
+            "-3836286746260530032892706.6174",
+            "1111481055212557787730018.3040",
+            Less,
+        ),
+        (
+            "3836286746260530032892706.6174",
+            "-1111481055212557787730018.3040",
+            Greater,
+        ),
+        (
+            "-3836286746260530032892706.6174",
+            "-1111481055212557787730018.3040",
+            Less,
+        ),
+        // 111 011 different scale
+        (
+            "79710135995301690627798250.27",
+            "4613684285077.479267304996864",
+            Greater,
+        ),
+        ("-79710135995301690627798250.27", "4613684285077.479267304996864", Less),
+        (
+            "79710135995301690627798250.27",
+            "-4613684285077.479267304996864",
+            Greater,
+        ),
+        ("-79710135995301690627798250.27", "-4613684285077.479267304996864", Less),
+        // 111 111 same scale
+        (
+            "3836286746260530032892706.6174",
+            "1881105048659612897896770.8539",
+            Greater,
+        ),
+        (
+            "-3836286746260530032892706.6174",
+            "1881105048659612897896770.8539",
+            Less,
+        ),
+        (
+            "3836286746260530032892706.6174",
+            "-1881105048659612897896770.8539",
+            Greater,
+        ),
+        (
+            "-3836286746260530032892706.6174",
+            "-1881105048659612897896770.8539",
+            Less,
+        ),
+        // 111 111 different scale
+        (
+            "3879592276836332218003.2886500",
+            "35612499407667292.686490959658",
+            Greater,
+        ),
+        (
+            "-3879592276836332218003.2886500",
+            "35612499407667292.686490959658",
+            Less,
+        ),
+        (
+            "3879592276836332218003.2886500",
+            "-35612499407667292.686490959658",
+            Greater,
+        ),
+        (
+            "-3879592276836332218003.2886500",
+            "-35612499407667292.686490959658",
+            Less,
+        ),
+    ];
+    for &(a, b, c) in tests {
+        cmp(a, b, c);
+    }
+}
+
+#[test]
+fn it_floors_decimals() {
+    let tests = &[
+        ("1", "1"),
+        ("1.00", "1"),
+        ("1.2345", "1"),
+        ("-1", "-1"),
+        ("-1.00", "-1"),
+        ("-1.2345", "-2"),
+    ];
+    for &(a, expected) in tests {
+        let a = Decimal::from_str(a).unwrap();
+        assert_eq!(expected, a.floor().to_string(), "Failed flooring {a}");
+    }
+}
+
+#[test]
+fn it_ceils_decimals() {
+    let tests = &[
+        ("1", "1"),
+        ("1.00", "1"),
+        ("1.2345", "2"),
+        ("-1", "-1"),
+        ("-1.00", "-1"),
+        ("-1.2345", "-1"),
+    ];
+    for &(a, expected) in tests {
+        let a = Decimal::from_str(a).unwrap();
+        assert_eq!(expected, a.ceil().to_string(), "Failed ceiling {a}");
+    }
+}
+
+#[test]
+fn it_finds_max_of_two() {
+    let tests = &[("1", "1", "1"), ("2", "1", "2"), ("1", "2", "2")];
+    for &(a, b, expected) in tests {
+        let a = Decimal::from_str(a).unwrap();
+        let b = Decimal::from_str(b).unwrap();
+        assert_eq!(expected, a.max(b).to_string());
+    }
+}
+
+#[test]
+fn it_finds_min_of_two() {
+    let tests = &[("1", "1", "1"), ("2", "1", "1"), ("1", "2", "1")];
+    for &(a, b, expected) in tests {
+        let a = Decimal::from_str(a).unwrap();
+        let b = Decimal::from_str(b).unwrap();
+        assert_eq!(expected, a.min(b).to_string());
+    }
+}
+
+#[test]
+fn test_max_compares() {
+    let x = "225.33543601344182".parse::<Decimal>().unwrap();
+    let y = Decimal::MAX;
+    assert!(x < y);
+    assert!(y > x);
+    assert_ne!(y, x);
+}
+
+#[test]
+fn test_min_compares() {
+    let x = "225.33543601344182".parse::<Decimal>().unwrap();
+    let y = Decimal::MIN;
+    assert!(x > y);
+    assert!(y < x);
+    assert_ne!(y, x);
+}
+
+#[test]
+fn test_zero_eq_negative_zero() {
+    let zero: Decimal = 0.into();
+
+    assert_eq!(zero, zero);
+    assert_eq!(-zero, zero);
+    assert_eq!(zero, -zero);
+}
+
+fn hash_it(d: Decimal) -> u64 {
+    use core::hash::{Hash, Hasher};
+    use std::collections::hash_map::DefaultHasher;
+
+    let mut h = DefaultHasher::new();
+    d.hash(&mut h);
+    h.finish()
+}
+
+#[test]
+fn it_computes_equal_hashes_for_equal_values() {
+    // From the Rust Hash docs:
+    //
+    // "When implementing both Hash and Eq, it is important that the following property holds:
+    //
+    //     k1 == k2 -> hash(k1) == hash(k2)"
+
+    let k1 = Decimal::from_str("1").unwrap();
+    let k2 = Decimal::from_str("1.0").unwrap();
+    let k3 = Decimal::from_str("1.00").unwrap();
+    let k4 = Decimal::from_str("1.01").unwrap();
+
+    assert_eq!(k1, k2);
+    assert_eq!(k1, k3);
+    assert_ne!(k1, k4);
+
+    let h1 = hash_it(k1);
+    let h2 = hash_it(k2);
+    let h3 = hash_it(k3);
+    let h4 = hash_it(k4);
+
+    assert_eq!(h1, h2);
+    assert_eq!(h1, h3);
+    assert_ne!(h1, h4);
+
+    // Test the application of Hash calculation to a HashMap.
+
+    let mut map = std::collections::HashMap::new();
+
+    map.insert(k1, k1.to_string());
+    // map[k2] should overwrite map[k1] because k1 == k2.
+    map.insert(k2, k2.to_string());
+
+    assert_eq!("1.0", map.get(&k3).expect("could not get k3"));
+    assert_eq!(1, map.len());
+
+    // map[k3] should overwrite map[k2] because k3 == k2.
+    map.insert(k3, k3.to_string());
+    // map[k4] should not overwrite map[k3] because k4 != k3.
+    map.insert(k4, k4.to_string());
+
+    assert_eq!(2, map.len());
+    assert_eq!("1.00", map.get(&k1).expect("could not get k1"));
+}
+
+#[test]
+fn it_computes_equal_hashes_for_positive_and_negative_zero() {
+    // Verify 0 and -0 have the same hash
+    let k1 = Decimal::from_str("0").unwrap();
+    let k2 = Decimal::from_str("-0").unwrap();
+    assert_eq!(k1, k2);
+    let h1 = hash_it(k1);
+    let h2 = hash_it(k2);
+    assert_eq!(h1, h2);
+
+    // Verify 0 and -0.0 have the same hash
+    let k1 = Decimal::from_str("0").unwrap();
+    let k2 = Decimal::from_str("-0.0").unwrap();
+    assert_eq!(k1, k2);
+    let h1 = hash_it(k1);
+    let h2 = hash_it(k2);
+    assert_eq!(h1, h2);
+}
