+++ tests/decimal_tests/main.rs
@@ -1,0 +1,11 @@
+mod arithmetic;
+mod comparison;
+mod conversions;
+mod formatting;
+mod issues;
+#[cfg(feature = "maths")]
+mod maths;
+mod misc;
+mod parsing;
+mod rounding;
+mod scale;
