+++ tests/generated.rs
@@ -1,0 +1,353 @@
+use rust_decimal::prelude::*;
+
+macro_rules! gen_test {
+    ($name:ident, $csv:expr, $method:tt) => {
+        #[test]
+        fn $name() {
+            let path = std::env::current_dir().unwrap();
+            let mut rdr = csv::Reader::from_reader(
+                std::fs::File::open(format!("{}/tests/generated/{}", path.display(), $csv)).unwrap(),
+            );
+            let mut row = 0;
+            for result in rdr.records() {
+                let record = result.unwrap();
+                row += 1;
+
+                // Extract the data
+                let d1 = record.get(0).unwrap();
+                let d2 = record.get(1).unwrap();
+                let result = record.get(2).unwrap();
+                let error = record.get(3).unwrap();
+
+                // Do the calc
+                let d1 = Decimal::from_str(&d1).unwrap();
+                let d2 = Decimal::from_str(&d2).unwrap();
+                let expected = Decimal::from_str(&result).unwrap();
+                match d1.$method(d2) {
+                    Some(v) => assert_eq!(expected, v, "Row {}", row),
+                    None => assert!(!error.is_empty()),
+                }
+            }
+        }
+    };
+}
+
+gen_test!(test_add_000_001, "Add_000_001.csv", checked_add);
+gen_test!(test_add_000_010, "Add_000_010.csv", checked_add);
+gen_test!(test_add_000_011, "Add_000_011.csv", checked_add);
+gen_test!(test_add_000_100, "Add_000_100.csv", checked_add);
+gen_test!(test_add_000_101, "Add_000_101.csv", checked_add);
+gen_test!(test_add_000_110, "Add_000_110.csv", checked_add);
+gen_test!(test_add_000_111, "Add_000_111.csv", checked_add);
+gen_test!(test_add_001_000, "Add_001_000.csv", checked_add);
+gen_test!(test_add_001_001, "Add_001_001.csv", checked_add);
+gen_test!(test_add_001_010, "Add_001_010.csv", checked_add);
+gen_test!(test_add_001_011, "Add_001_011.csv", checked_add);
+gen_test!(test_add_001_100, "Add_001_100.csv", checked_add);
+gen_test!(test_add_001_101, "Add_001_101.csv", checked_add);
+gen_test!(test_add_001_110, "Add_001_110.csv", checked_add);
+gen_test!(test_add_001_111, "Add_001_111.csv", checked_add);
+gen_test!(test_add_010_000, "Add_010_000.csv", checked_add);
+gen_test!(test_add_010_001, "Add_010_001.csv", checked_add);
+gen_test!(test_add_010_010, "Add_010_010.csv", checked_add);
+gen_test!(test_add_010_011, "Add_010_011.csv", checked_add);
+gen_test!(test_add_010_100, "Add_010_100.csv", checked_add);
+gen_test!(test_add_010_101, "Add_010_101.csv", checked_add);
+gen_test!(test_add_010_110, "Add_010_110.csv", checked_add);
+gen_test!(test_add_010_111, "Add_010_111.csv", checked_add);
+gen_test!(test_add_011_000, "Add_011_000.csv", checked_add);
+gen_test!(test_add_011_001, "Add_011_001.csv", checked_add);
+gen_test!(test_add_011_010, "Add_011_010.csv", checked_add);
+gen_test!(test_add_011_011, "Add_011_011.csv", checked_add);
+gen_test!(test_add_011_100, "Add_011_100.csv", checked_add);
+gen_test!(test_add_011_101, "Add_011_101.csv", checked_add);
+gen_test!(test_add_011_110, "Add_011_110.csv", checked_add);
+gen_test!(test_add_011_111, "Add_011_111.csv", checked_add);
+gen_test!(test_add_100_000, "Add_100_000.csv", checked_add);
+gen_test!(test_add_100_001, "Add_100_001.csv", checked_add);
+gen_test!(test_add_100_010, "Add_100_010.csv", checked_add);
+gen_test!(test_add_100_011, "Add_100_011.csv", checked_add);
+gen_test!(test_add_100_100, "Add_100_100.csv", checked_add);
+gen_test!(test_add_100_101, "Add_100_101.csv", checked_add);
+gen_test!(test_add_100_110, "Add_100_110.csv", checked_add);
+gen_test!(test_add_100_111, "Add_100_111.csv", checked_add);
+gen_test!(test_add_101_000, "Add_101_000.csv", checked_add);
+gen_test!(test_add_101_001, "Add_101_001.csv", checked_add);
+gen_test!(test_add_101_010, "Add_101_010.csv", checked_add);
+gen_test!(test_add_101_011, "Add_101_011.csv", checked_add);
+gen_test!(test_add_101_100, "Add_101_100.csv", checked_add);
+gen_test!(test_add_101_101, "Add_101_101.csv", checked_add);
+gen_test!(test_add_101_110, "Add_101_110.csv", checked_add);
+gen_test!(test_add_101_111, "Add_101_111.csv", checked_add);
+gen_test!(test_add_110_000, "Add_110_000.csv", checked_add);
+gen_test!(test_add_110_001, "Add_110_001.csv", checked_add);
+gen_test!(test_add_110_010, "Add_110_010.csv", checked_add);
+gen_test!(test_add_110_011, "Add_110_011.csv", checked_add);
+gen_test!(test_add_110_100, "Add_110_100.csv", checked_add);
+gen_test!(test_add_110_101, "Add_110_101.csv", checked_add);
+gen_test!(test_add_110_110, "Add_110_110.csv", checked_add);
+gen_test!(test_add_110_111, "Add_110_111.csv", checked_add);
+gen_test!(test_add_111_000, "Add_111_000.csv", checked_add);
+gen_test!(test_add_111_001, "Add_111_001.csv", checked_add);
+gen_test!(test_add_111_010, "Add_111_010.csv", checked_add);
+gen_test!(test_add_111_011, "Add_111_011.csv", checked_add);
+gen_test!(test_add_111_100, "Add_111_100.csv", checked_add);
+gen_test!(test_add_111_101, "Add_111_101.csv", checked_add);
+gen_test!(test_add_111_110, "Add_111_110.csv", checked_add);
+gen_test!(test_add_111_111, "Add_111_111.csv", checked_add);
+
+gen_test!(test_div_000_001, "Div_000_001.csv", checked_div);
+gen_test!(test_div_000_010, "Div_000_010.csv", checked_div);
+gen_test!(test_div_000_011, "Div_000_011.csv", checked_div);
+gen_test!(test_div_000_100, "Div_000_100.csv", checked_div);
+gen_test!(test_div_000_101, "Div_000_101.csv", checked_div);
+gen_test!(test_div_000_110, "Div_000_110.csv", checked_div);
+gen_test!(test_div_000_111, "Div_000_111.csv", checked_div);
+gen_test!(test_div_001_000, "Div_001_000.csv", checked_div);
+gen_test!(test_div_001_001, "Div_001_001.csv", checked_div);
+gen_test!(test_div_001_010, "Div_001_010.csv", checked_div);
+gen_test!(test_div_001_011, "Div_001_011.csv", checked_div);
+gen_test!(test_div_001_100, "Div_001_100.csv", checked_div);
+gen_test!(test_div_001_101, "Div_001_101.csv", checked_div);
+gen_test!(test_div_001_110, "Div_001_110.csv", checked_div);
+gen_test!(test_div_001_111, "Div_001_111.csv", checked_div);
+gen_test!(test_div_010_000, "Div_010_000.csv", checked_div);
+gen_test!(test_div_010_001, "Div_010_001.csv", checked_div);
+gen_test!(test_div_010_010, "Div_010_010.csv", checked_div);
+gen_test!(test_div_010_011, "Div_010_011.csv", checked_div);
+gen_test!(test_div_010_100, "Div_010_100.csv", checked_div);
+gen_test!(test_div_010_101, "Div_010_101.csv", checked_div);
+gen_test!(test_div_010_110, "Div_010_110.csv", checked_div);
+gen_test!(test_div_010_111, "Div_010_111.csv", checked_div);
+gen_test!(test_div_011_000, "Div_011_000.csv", checked_div);
+gen_test!(test_div_011_001, "Div_011_001.csv", checked_div);
+gen_test!(test_div_011_010, "Div_011_010.csv", checked_div);
+gen_test!(test_div_011_011, "Div_011_011.csv", checked_div);
+gen_test!(test_div_011_100, "Div_011_100.csv", checked_div);
+gen_test!(test_div_011_101, "Div_011_101.csv", checked_div);
+gen_test!(test_div_011_110, "Div_011_110.csv", checked_div);
+gen_test!(test_div_011_111, "Div_011_111.csv", checked_div);
+gen_test!(test_div_100_000, "Div_100_000.csv", checked_div);
+gen_test!(test_div_100_001, "Div_100_001.csv", checked_div);
+gen_test!(test_div_100_010, "Div_100_010.csv", checked_div);
+gen_test!(test_div_100_011, "Div_100_011.csv", checked_div);
+gen_test!(test_div_100_100, "Div_100_100.csv", checked_div);
+gen_test!(test_div_100_101, "Div_100_101.csv", checked_div);
+gen_test!(test_div_100_110, "Div_100_110.csv", checked_div);
+gen_test!(test_div_100_111, "Div_100_111.csv", checked_div);
+gen_test!(test_div_101_000, "Div_101_000.csv", checked_div);
+gen_test!(test_div_101_001, "Div_101_001.csv", checked_div);
+gen_test!(test_div_101_010, "Div_101_010.csv", checked_div);
+gen_test!(test_div_101_011, "Div_101_011.csv", checked_div);
+gen_test!(test_div_101_100, "Div_101_100.csv", checked_div);
+gen_test!(test_div_101_101, "Div_101_101.csv", checked_div);
+gen_test!(test_div_101_110, "Div_101_110.csv", checked_div);
+gen_test!(test_div_101_111, "Div_101_111.csv", checked_div);
+gen_test!(test_div_110_000, "Div_110_000.csv", checked_div);
+gen_test!(test_div_110_001, "Div_110_001.csv", checked_div);
+gen_test!(test_div_110_010, "Div_110_010.csv", checked_div);
+gen_test!(test_div_110_011, "Div_110_011.csv", checked_div);
+gen_test!(test_div_110_100, "Div_110_100.csv", checked_div);
+gen_test!(test_div_110_101, "Div_110_101.csv", checked_div);
+gen_test!(test_div_110_110, "Div_110_110.csv", checked_div);
+gen_test!(test_div_110_111, "Div_110_111.csv", checked_div);
+gen_test!(test_div_111_000, "Div_111_000.csv", checked_div);
+gen_test!(test_div_111_001, "Div_111_001.csv", checked_div);
+gen_test!(test_div_111_010, "Div_111_010.csv", checked_div);
+gen_test!(test_div_111_011, "Div_111_011.csv", checked_div);
+gen_test!(test_div_111_100, "Div_111_100.csv", checked_div);
+gen_test!(test_div_111_101, "Div_111_101.csv", checked_div);
+gen_test!(test_div_111_110, "Div_111_110.csv", checked_div);
+gen_test!(test_div_111_111, "Div_111_111.csv", checked_div);
+
+gen_test!(test_mul_000_001, "Mul_000_001.csv", checked_mul);
+gen_test!(test_mul_000_010, "Mul_000_010.csv", checked_mul);
+gen_test!(test_mul_000_011, "Mul_000_011.csv", checked_mul);
+gen_test!(test_mul_000_100, "Mul_000_100.csv", checked_mul);
+gen_test!(test_mul_000_101, "Mul_000_101.csv", checked_mul);
+gen_test!(test_mul_000_110, "Mul_000_110.csv", checked_mul);
+gen_test!(test_mul_000_111, "Mul_000_111.csv", checked_mul);
+gen_test!(test_mul_001_000, "Mul_001_000.csv", checked_mul);
+gen_test!(test_mul_001_001, "Mul_001_001.csv", checked_mul);
+gen_test!(test_mul_001_010, "Mul_001_010.csv", checked_mul);
+gen_test!(test_mul_001_011, "Mul_001_011.csv", checked_mul);
+gen_test!(test_mul_001_100, "Mul_001_100.csv", checked_mul);
+gen_test!(test_mul_001_101, "Mul_001_101.csv", checked_mul);
+gen_test!(test_mul_001_110, "Mul_001_110.csv", checked_mul);
+gen_test!(test_mul_001_111, "Mul_001_111.csv", checked_mul);
+gen_test!(test_mul_010_000, "Mul_010_000.csv", checked_mul);
+gen_test!(test_mul_010_001, "Mul_010_001.csv", checked_mul);
+gen_test!(test_mul_010_010, "Mul_010_010.csv", checked_mul);
+gen_test!(test_mul_010_011, "Mul_010_011.csv", checked_mul);
+gen_test!(test_mul_010_100, "Mul_010_100.csv", checked_mul);
+gen_test!(test_mul_010_101, "Mul_010_101.csv", checked_mul);
+gen_test!(test_mul_010_110, "Mul_010_110.csv", checked_mul);
+gen_test!(test_mul_010_111, "Mul_010_111.csv", checked_mul);
+gen_test!(test_mul_011_000, "Mul_011_000.csv", checked_mul);
+gen_test!(test_mul_011_001, "Mul_011_001.csv", checked_mul);
+gen_test!(test_mul_011_010, "Mul_011_010.csv", checked_mul);
+gen_test!(test_mul_011_011, "Mul_011_011.csv", checked_mul);
+gen_test!(test_mul_011_100, "Mul_011_100.csv", checked_mul);
+gen_test!(test_mul_011_101, "Mul_011_101.csv", checked_mul);
+gen_test!(test_mul_011_110, "Mul_011_110.csv", checked_mul);
+gen_test!(test_mul_011_111, "Mul_011_111.csv", checked_mul);
+gen_test!(test_mul_100_000, "Mul_100_000.csv", checked_mul);
+gen_test!(test_mul_100_001, "Mul_100_001.csv", checked_mul);
+gen_test!(test_mul_100_010, "Mul_100_010.csv", checked_mul);
+gen_test!(test_mul_100_011, "Mul_100_011.csv", checked_mul);
+gen_test!(test_mul_100_100, "Mul_100_100.csv", checked_mul);
+gen_test!(test_mul_100_101, "Mul_100_101.csv", checked_mul);
+gen_test!(test_mul_100_110, "Mul_100_110.csv", checked_mul);
+gen_test!(test_mul_100_111, "Mul_100_111.csv", checked_mul);
+gen_test!(test_mul_101_000, "Mul_101_000.csv", checked_mul);
+gen_test!(test_mul_101_001, "Mul_101_001.csv", checked_mul);
+gen_test!(test_mul_101_010, "Mul_101_010.csv", checked_mul);
+gen_test!(test_mul_101_011, "Mul_101_011.csv", checked_mul);
+gen_test!(test_mul_101_100, "Mul_101_100.csv", checked_mul);
+gen_test!(test_mul_101_101, "Mul_101_101.csv", checked_mul);
+gen_test!(test_mul_101_110, "Mul_101_110.csv", checked_mul);
+gen_test!(test_mul_101_111, "Mul_101_111.csv", checked_mul);
+gen_test!(test_mul_110_000, "Mul_110_000.csv", checked_mul);
+gen_test!(test_mul_110_001, "Mul_110_001.csv", checked_mul);
+gen_test!(test_mul_110_010, "Mul_110_010.csv", checked_mul);
+gen_test!(test_mul_110_011, "Mul_110_011.csv", checked_mul);
+gen_test!(test_mul_110_100, "Mul_110_100.csv", checked_mul);
+gen_test!(test_mul_110_101, "Mul_110_101.csv", checked_mul);
+gen_test!(test_mul_110_110, "Mul_110_110.csv", checked_mul);
+gen_test!(test_mul_110_111, "Mul_110_111.csv", checked_mul);
+gen_test!(test_mul_111_000, "Mul_111_000.csv", checked_mul);
+gen_test!(test_mul_111_001, "Mul_111_001.csv", checked_mul);
+gen_test!(test_mul_111_010, "Mul_111_010.csv", checked_mul);
+gen_test!(test_mul_111_011, "Mul_111_011.csv", checked_mul);
+gen_test!(test_mul_111_100, "Mul_111_100.csv", checked_mul);
+gen_test!(test_mul_111_101, "Mul_111_101.csv", checked_mul);
+gen_test!(test_mul_111_110, "Mul_111_110.csv", checked_mul);
+gen_test!(test_mul_111_111, "Mul_111_111.csv", checked_mul);
+
+gen_test!(test_rem_000_001, "Rem_000_001.csv", checked_rem);
+gen_test!(test_rem_000_010, "Rem_000_010.csv", checked_rem);
+gen_test!(test_rem_000_011, "Rem_000_011.csv", checked_rem);
+gen_test!(test_rem_000_100, "Rem_000_100.csv", checked_rem);
+gen_test!(test_rem_000_101, "Rem_000_101.csv", checked_rem);
+gen_test!(test_rem_000_110, "Rem_000_110.csv", checked_rem);
+gen_test!(test_rem_000_111, "Rem_000_111.csv", checked_rem);
+gen_test!(test_rem_001_000, "Rem_001_000.csv", checked_rem);
+gen_test!(test_rem_001_001, "Rem_001_001.csv", checked_rem);
+gen_test!(test_rem_001_010, "Rem_001_010.csv", checked_rem);
+gen_test!(test_rem_001_011, "Rem_001_011.csv", checked_rem);
+gen_test!(test_rem_001_100, "Rem_001_100.csv", checked_rem);
+gen_test!(test_rem_001_101, "Rem_001_101.csv", checked_rem);
+gen_test!(test_rem_001_110, "Rem_001_110.csv", checked_rem);
+gen_test!(test_rem_001_111, "Rem_001_111.csv", checked_rem);
+gen_test!(test_rem_010_000, "Rem_010_000.csv", checked_rem);
+gen_test!(test_rem_010_001, "Rem_010_001.csv", checked_rem);
+gen_test!(test_rem_010_010, "Rem_010_010.csv", checked_rem);
+gen_test!(test_rem_010_011, "Rem_010_011.csv", checked_rem);
+gen_test!(test_rem_010_100, "Rem_010_100.csv", checked_rem);
+gen_test!(test_rem_010_101, "Rem_010_101.csv", checked_rem);
+gen_test!(test_rem_010_110, "Rem_010_110.csv", checked_rem);
+gen_test!(test_rem_010_111, "Rem_010_111.csv", checked_rem);
+gen_test!(test_rem_011_000, "Rem_011_000.csv", checked_rem);
+gen_test!(test_rem_011_001, "Rem_011_001.csv", checked_rem);
+gen_test!(test_rem_011_010, "Rem_011_010.csv", checked_rem);
+gen_test!(test_rem_011_011, "Rem_011_011.csv", checked_rem);
+gen_test!(test_rem_011_100, "Rem_011_100.csv", checked_rem);
+gen_test!(test_rem_011_101, "Rem_011_101.csv", checked_rem);
+gen_test!(test_rem_011_110, "Rem_011_110.csv", checked_rem);
+gen_test!(test_rem_011_111, "Rem_011_111.csv", checked_rem);
+gen_test!(test_rem_100_000, "Rem_100_000.csv", checked_rem);
+gen_test!(test_rem_100_001, "Rem_100_001.csv", checked_rem);
+gen_test!(test_rem_100_010, "Rem_100_010.csv", checked_rem);
+gen_test!(test_rem_100_011, "Rem_100_011.csv", checked_rem);
+gen_test!(test_rem_100_100, "Rem_100_100.csv", checked_rem);
+gen_test!(test_rem_100_101, "Rem_100_101.csv", checked_rem);
+gen_test!(test_rem_100_110, "Rem_100_110.csv", checked_rem);
+gen_test!(test_rem_100_111, "Rem_100_111.csv", checked_rem);
+gen_test!(test_rem_101_000, "Rem_101_000.csv", checked_rem);
+gen_test!(test_rem_101_001, "Rem_101_001.csv", checked_rem);
+gen_test!(test_rem_101_010, "Rem_101_010.csv", checked_rem);
+gen_test!(test_rem_101_011, "Rem_101_011.csv", checked_rem);
+gen_test!(test_rem_101_100, "Rem_101_100.csv", checked_rem);
+gen_test!(test_rem_101_101, "Rem_101_101.csv", checked_rem);
+gen_test!(test_rem_101_110, "Rem_101_110.csv", checked_rem);
+gen_test!(test_rem_101_111, "Rem_101_111.csv", checked_rem);
+gen_test!(test_rem_110_000, "Rem_110_000.csv", checked_rem);
+gen_test!(test_rem_110_001, "Rem_110_001.csv", checked_rem);
+gen_test!(test_rem_110_010, "Rem_110_010.csv", checked_rem);
+gen_test!(test_rem_110_011, "Rem_110_011.csv", checked_rem);
+gen_test!(test_rem_110_100, "Rem_110_100.csv", checked_rem);
+gen_test!(test_rem_110_101, "Rem_110_101.csv", checked_rem);
+gen_test!(test_rem_110_110, "Rem_110_110.csv", checked_rem);
+gen_test!(test_rem_110_111, "Rem_110_111.csv", checked_rem);
+gen_test!(test_rem_111_000, "Rem_111_000.csv", checked_rem);
+gen_test!(test_rem_111_001, "Rem_111_001.csv", checked_rem);
+gen_test!(test_rem_111_010, "Rem_111_010.csv", checked_rem);
+gen_test!(test_rem_111_011, "Rem_111_011.csv", checked_rem);
+gen_test!(test_rem_111_100, "Rem_111_100.csv", checked_rem);
+gen_test!(test_rem_111_101, "Rem_111_101.csv", checked_rem);
+gen_test!(test_rem_111_110, "Rem_111_110.csv", checked_rem);
+gen_test!(test_rem_111_111, "Rem_111_111.csv", checked_rem);
+
+gen_test!(test_sub_000_001, "Sub_000_001.csv", checked_sub);
+gen_test!(test_sub_000_010, "Sub_000_010.csv", checked_sub);
+gen_test!(test_sub_000_011, "Sub_000_011.csv", checked_sub);
+gen_test!(test_sub_000_100, "Sub_000_100.csv", checked_sub);
+gen_test!(test_sub_000_101, "Sub_000_101.csv", checked_sub);
+gen_test!(test_sub_000_110, "Sub_000_110.csv", checked_sub);
+gen_test!(test_sub_000_111, "Sub_000_111.csv", checked_sub);
+gen_test!(test_sub_001_000, "Sub_001_000.csv", checked_sub);
+gen_test!(test_sub_001_001, "Sub_001_001.csv", checked_sub);
+gen_test!(test_sub_001_010, "Sub_001_010.csv", checked_sub);
+gen_test!(test_sub_001_011, "Sub_001_011.csv", checked_sub);
+gen_test!(test_sub_001_100, "Sub_001_100.csv", checked_sub);
+gen_test!(test_sub_001_101, "Sub_001_101.csv", checked_sub);
+gen_test!(test_sub_001_110, "Sub_001_110.csv", checked_sub);
+gen_test!(test_sub_001_111, "Sub_001_111.csv", checked_sub);
+gen_test!(test_sub_010_000, "Sub_010_000.csv", checked_sub);
+gen_test!(test_sub_010_001, "Sub_010_001.csv", checked_sub);
+gen_test!(test_sub_010_010, "Sub_010_010.csv", checked_sub);
+gen_test!(test_sub_010_011, "Sub_010_011.csv", checked_sub);
+gen_test!(test_sub_010_100, "Sub_010_100.csv", checked_sub);
+gen_test!(test_sub_010_101, "Sub_010_101.csv", checked_sub);
+gen_test!(test_sub_010_110, "Sub_010_110.csv", checked_sub);
+gen_test!(test_sub_010_111, "Sub_010_111.csv", checked_sub);
+gen_test!(test_sub_011_000, "Sub_011_000.csv", checked_sub);
+gen_test!(test_sub_011_001, "Sub_011_001.csv", checked_sub);
+gen_test!(test_sub_011_010, "Sub_011_010.csv", checked_sub);
+gen_test!(test_sub_011_011, "Sub_011_011.csv", checked_sub);
+gen_test!(test_sub_011_100, "Sub_011_100.csv", checked_sub);
+gen_test!(test_sub_011_101, "Sub_011_101.csv", checked_sub);
+gen_test!(test_sub_011_110, "Sub_011_110.csv", checked_sub);
+gen_test!(test_sub_011_111, "Sub_011_111.csv", checked_sub);
+gen_test!(test_sub_100_000, "Sub_100_000.csv", checked_sub);
+gen_test!(test_sub_100_001, "Sub_100_001.csv", checked_sub);
+gen_test!(test_sub_100_010, "Sub_100_010.csv", checked_sub);
+gen_test!(test_sub_100_011, "Sub_100_011.csv", checked_sub);
+gen_test!(test_sub_100_100, "Sub_100_100.csv", checked_sub);
+gen_test!(test_sub_100_101, "Sub_100_101.csv", checked_sub);
+gen_test!(test_sub_100_110, "Sub_100_110.csv", checked_sub);
+gen_test!(test_sub_100_111, "Sub_100_111.csv", checked_sub);
+gen_test!(test_sub_101_000, "Sub_101_000.csv", checked_sub);
+gen_test!(test_sub_101_001, "Sub_101_001.csv", checked_sub);
+gen_test!(test_sub_101_010, "Sub_101_010.csv", checked_sub);
+gen_test!(test_sub_101_011, "Sub_101_011.csv", checked_sub);
+gen_test!(test_sub_101_100, "Sub_101_100.csv", checked_sub);
+gen_test!(test_sub_101_101, "Sub_101_101.csv", checked_sub);
+gen_test!(test_sub_101_110, "Sub_101_110.csv", checked_sub);
+gen_test!(test_sub_101_111, "Sub_101_111.csv", checked_sub);
+gen_test!(test_sub_110_000, "Sub_110_000.csv", checked_sub);
+gen_test!(test_sub_110_001, "Sub_110_001.csv", checked_sub);
+gen_test!(test_sub_110_010, "Sub_110_010.csv", checked_sub);
+gen_test!(test_sub_110_011, "Sub_110_011.csv", checked_sub);
+gen_test!(test_sub_110_100, "Sub_110_100.csv", checked_sub);
+gen_test!(test_sub_110_101, "Sub_110_101.csv", checked_sub);
+gen_test!(test_sub_110_110, "Sub_110_110.csv", checked_sub);
+gen_test!(test_sub_110_111, "Sub_110_111.csv", checked_sub);
+gen_test!(test_sub_111_000, "Sub_111_000.csv", checked_sub);
+gen_test!(test_sub_111_001, "Sub_111_001.csv", checked_sub);
+gen_test!(test_sub_111_010, "Sub_111_010.csv", checked_sub);
+gen_test!(test_sub_111_011, "Sub_111_011.csv", checked_sub);
+gen_test!(test_sub_111_100, "Sub_111_100.csv", checked_sub);
+gen_test!(test_sub_111_101, "Sub_111_101.csv", checked_sub);
+gen_test!(test_sub_111_110, "Sub_111_110.csv", checked_sub);
+gen_test!(test_sub_111_111, "Sub_111_111.csv", checked_sub);
